requirements.txt
streamlit
google-generativeai
python-dotenv
ipython
streamlit_extras
